REFACTOR Seperated deletion of messages into seperate class
